requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
opencv-python==4.5.3.56
pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
pandas==1.3.0